Move funcs to an extra file
