+++ config/configload/load.go
@@ -2,14 +2,12 @@ package configload
 
 import (
 	"fmt"
-	"os"
 	"path"
 	"path/filepath"
 	"strings"
 
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/gohcl"
-	"github.com/hashicorp/hcl/v2/hclparse"
 	"github.com/hashicorp/hcl/v2/hclsyntax"
 	"github.com/zclconf/go-cty/cty"
 
@@ -85,49 +83,6 @@ func updateContext(body hcl.Body, srcBytes [][]byte, environment string) hcl.Dia
 	envContext = evalContext.HCLContext() // global assign
 
 	return nil
-}
-
-func parseFile(filePath string, srcBytes *[][]byte) (*hclsyntax.Body, error) {
-	src, err := os.ReadFile(filePath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to load configuration: %w", err)
-	}
-
-	*srcBytes = append(*srcBytes, src)
-
-	parsed, diags := hclparse.NewParser().ParseHCLFile(filePath)
-	if diags.HasErrors() {
-		return nil, diags
-	}
-
-	return parsed.Body.(*hclsyntax.Body), nil
-}
-
-func parseFiles(files configfile.Files) ([]*hclsyntax.Body, [][]byte, error) {
-	var (
-		srcBytes     [][]byte
-		parsedBodies []*hclsyntax.Body
-	)
-
-	for _, file := range files {
-		if file.IsDir {
-			childBodies, bytes, err := parseFiles(file.Children)
-			if err != nil {
-				return nil, bytes, err
-			}
-
-			parsedBodies = append(parsedBodies, childBodies...)
-			srcBytes = append(srcBytes, bytes...)
-		} else {
-			body, err := parseFile(file.Path, &srcBytes)
-			if err != nil {
-				return nil, srcBytes, err
-			}
-			parsedBodies = append(parsedBodies, body)
-		}
-	}
-
-	return parsedBodies, srcBytes, nil
 }
 
 func bodiesToConfig(parsedBodies []*hclsyntax.Body, srcBytes [][]byte, env string) (*config.Couper, error) {
